requirements.txt
torch
ultralytics
supervision
pillow
numpy
tqdm
matplotlib
requests
